hubspot_ticket_automation/scripts/requirements.txt
google-auth>=2.23
google-auth-oauthlib>=1.2
google-api-python-client>=2.100
PyYAML>=6.0
